ResearchAssistant/Deployment/scripts/aihub_scripts/requirements.txt
azure-identity
azure-keyvault-secrets
azure-ai-resources
azure-ai-ml==1.17.1
msal[broker]==1.24.0b1
azure-mgmt-resource==23.1.0b2
azure-mgmt-cognitiveservices
azure-mgmt-search
azure-mgmt-storage
azure-mgmt-machinelearningservices